+++ README.md
@@ -1,0 +1,37 @@
+# BigKeeper
+
+big-keeper is an efficiency improvement for iOS modular development, iOSer using this tool can make modular development easier.
+
+## Installation
+
+Add this line to your application's Gemfile:
+
+```ruby
+gem 'big_keeper'
+```
+
+And then execute:
+
+    $ bundle
+
+Or install it yourself as:
+
+    $ gem install big_keeper
+
+## Usage
+
+Run `big-keeper` to learn how to use big-stash.
+
+## Development
+
+After checking out the repo, run `bin/setup` to install dependencies. You can also run `bin/console` for an interactive prompt that will allow you to experiment.
+
+To install this gem onto your local machine, run `bundle exec rake install`. To release a new version, update the version number in `version.rb`, and then run `bundle exec rake release`, which will create a git tag for the version, push git commits and tags, and push the `.gem` file to [rubygems.org](https://rubygems.org).
+
+## Contributing
+
+Bug reports and pull requests are welcome on GitHub at https://github.com/BigKeeper/big-keeper.
+
+## License
+
+The gem is available as open source under the terms of the [MIT License](http://opensource.org/licenses/MIT).
