+++ lib/big_keeper.rb
@@ -1,0 +1,5 @@
+require "big_keeper/version"
+
+module BigKeeper
+  # Your code goes here...
+end
