+++ lib/big_keeper/command/start_module_release.sh
@@ -42,4 +42,3 @@ git tag -m "update podspec" $thisTag
 git push --tags
 
 pod repo push specs $podSpecName.podspec --allow-warnings
-
