+++ lib/big_keeper/version.rb
@@ -1,0 +1,3 @@
+module BigKeeper
+  VERSION = "0.1.0"
+end
