--- util/gitflow_operator.rb
@@ -1,13 +1,0 @@
-
-# Operator for gitflow
-class GitflowOperator
-  def create_feature(path, feature_name)
-    p `cd #{path}; git flow feature start #{feature_name}`
-  end
-
-  def stash(path, feature_name)
-  end
-
-  def apply_stash(path, feature_name)
-  end
-end
